+++ src/app/recipe/[id]/page.tsx
@@ -1,31 +1,39 @@
-"use client";
-
 import { myApi } from "@/app/api/instance";
-import RecipeDetailCard from "@/components/RecipeDetailCard";
+import { SITE } from "@/constants/env";
 import { ParamsWithIdI } from "@/interface/page";
 import { RecipeI } from "@/interface/recipe";
-import { RecipeDataForCard } from "@/utils/transform";
-import { useEffect, useState } from "react";
+import { Metadata } from "next";
+import RecipePage from "./RecipePage";
 
-export default function RecipePage({ params }: ParamsWithIdI) {
-  const [data, setData] = useState<RecipeI>();
+export async function generateMetadata({
+  params,
+}: ParamsWithIdI): Promise<Metadata> {
+  const { data } = await myApi.get<RecipeI>(`/recipe/${params.id}`);
 
-  async function fetchFeed() {
-    const res = await myApi.get<RecipeI>(`/recipe/${params.id}`);
-    setData(res.data);
-  }
+  const titleString = `필터레시피 | ${data.title}`;
+  const descriptionString = `#${data.category.main} #${data.category.sub} ${data.description}`;
 
-  useEffect(() => {
-    fetchFeed();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  return {
+    openGraph: {
+      title: titleString,
+      description: descriptionString,
+      url: `${SITE}/recipe/${params.id}`,
+      images: [
+        {
+          url: data.image.after,
+          width: 800,
+          height: 600,
+        },
+        {
+          url: data.image.before,
+          width: 800,
+          height: 600,
+        },
+      ],
+    },
+  };
+}
 
-  if (!data) return <div aria-busy="true" />;
-
-  const cardData = RecipeDataForCard(data);
-  return (
-    <div>
-      <RecipeDetailCard data={cardData} open={true} />
-    </div>
-  );
+export default function Page({ params }: ParamsWithIdI) {
+  return <RecipePage params={params} />;
 }
